+++ drizzled/function/locate.cc
@@ -50,7 +50,7 @@ int64_t Item_func_locate::val_int()
   {
     start0= start= args[2]->val_int() - 1;
 
-    if ((start < 0) || (start > a->length()))
+    if ((start < 0) || (start > static_cast<int64_t>(a->length())))
       return 0;
 
     /* start is now sufficiently valid to pass to charpos function */
